Added support for viewing pointers. Added link to Solarized. Removed Assembly File Version.
Pointers can now be found and parsed, but are not interactive yet. The
pointer is automatically shown by reversing the bytes and cropping the
0x08, making it easier to see and search for the address.

Solarized is a very useful color scheme - an about menu item was added
to link to the webpage.

The AssemblyInfo contained redundant information: the Assembly File
Version can be automatically generated from the Assembly Version, so it
was removed.

Refactor: the Font and ByteFlyweights were moved into the Utility class.