Update pointer visuals
I shouldn't hide data from the user: swapping the byte order makes the
address easier to see, but it hides the underlying bytes. The format has
been changed to make pointers still easily visible, but hopefully
without being as jarring.

Pointes have been changed from Red to Orange, so Red can be reserved
either as an error color, or as a "changed/not saved" color, as seen in
other hex editors.