First revision of visual interpretation
Images can now be viewed visually. Whenever any bytes for a compressed
image comes onscreen, that image is shown with a default dump palette,
at its default size. The width and height are guessed.

Future revisions may involve manually setting width/height, showing the
image on pointers to the data, changing between 16bit and 32bit,
ordering interpretations to appear in the same order as the data, and
highlighting the part of the data interpreted to create the image.