Reverse Pointers with multiple sources
Implement context menu for back pointers with multiple sources. If there
is more than one pointer to a location, a context menu will open with
all the possible sources. Selecting a source bring you to it.

There seems to be something wrong with the context menu style - that
will be fixed later.