Bugfixes for interpretation connect
In the case of pointers, the elements connected to an interpretation may
not all be in a single set: there may be multiple pointers, with space
between, pointing to the same data. The hover connect should check the
index of all elements.

Wrapped elements don't have a grid position: check their parents for a
grid position instead.