Increase Pointer Accuracy
Before, the app assumed that the first match for a pointer format is
always a pointer. Now, the app examines all possible pointer start
locations and uses a confidence metric to determine which is most likely
to be the real pointer.

- Pointers won't point to the inside of other pointers.
- Pointers won't point to a single byte of data.
- Pointers don't overlap other pointers.

Adding these three metrics removes several false positives and presents
less false negatives.