improve string parser
The string parser now caches interpretations after the first time
they're requested. This speeds up performance, removes a few bugs, and
prevents degredation of startup performance by not pre-computing them
all.

The parser now recognized strings that start with a single letter and a
space, such as "A ..." or "I ...".