Delay Load
Startup time is getting to be a bit slow, now that there are several
parsers running through the data and performing computations. Since this
work mostly involves building indexes of some sort, it can be offloaded
to another thread.

ElementFactory objects now have a Load method. They can choose to move
any non-STA thread work into Load, which can be done after the window is
shown. Once the load is complete, the window will use a Jump back to the
current location to refresh the window and interpretations.