outline for a thumbnail parser/interpreter
started an outline for parsing/interpreting thumbnail data. Would finish
it, except the 138, 13C, and 140 offsets are only true for 3 of the 5
roms - Ruby/Sapphire too different. Either the thumbnail parser needs
code specific to the two formats, or a new approach must be found.
also fixed a couple strange index-bounds bugs when running.