Adding Breadcrumb Bar
When searching forward/back through pointers, it is often useful to
return to where you just were - perhaps you followed the wrong pointer,
or it didn't lead you where you expected. You could Ctrl+Click on the
pointer, but this doesn't work if you don't remember which of the
pointers you came from. It's also jarring because it puts the pointer
back in the offset location, which may not be where it was. Also, you
may not have used a pointer - you may've jumped using a Goto.

A Breadcrumb bar has been added to assist in going backwards through the
data. It shows whenever the Find/Goto dialog doesn't, so it doesn't
require any additional real-estate.