Refactor
Trying to work on performance. Canvas is not perceivably  faster than a
Grid, but moving away from using ColumnDefinitionCount and
RowDefinitionCount towards currentColumnCount and currentRowCount is
still probably a useful change. Temporarily adding the canvas code as
comments, so I can switch back if needed.

Canvas elements don't auto-size, so the layout looks a little different.
That can be fixed if I provide manual sizes to the elements. But if it's
not faster, I won't add more code.