Map parser update
Now that I can parse all the elements correctly, I need a list of all
the subsections, and the ability to recognize them for pointers.

Also includes a pointer parser update: most pointers are found on word
boundaries and most point to word boundaries. Include this information
in the confidence estimation.