Refactor
As I begin working on the parsing and visualization of the maps, I
realized I'm going to need a smarter parser, capable of dealing with
multiple levels of data. A new parser has been created as an experiment
that may be able to replace multiple parsers with just this one. It may
also make parsing faster, as all the there will be less passes, and thus
less searches, when parsing.

The idea is to abstract out a few common differences in the
visualizations, and then allow other parsers to pass their results into
a single hub location. This could distill several parsers down to just
their load methods. Right now, the parser has only generic information,
so it's suitable for living outside the Gba namespace. If it eventually
requires pointer support (likely), the pointers will need to be stored
separately to keep the class outside the Gba namespace.

Also, since I've started thinking of the ElementFactories as Parsers,
I've gone ahead and renamed them. They parse the bytes and return
elements, so both names are appropriate.