Adding a string decoder based on my theory
It's possible to run the app when you know what strings are in the ROM,
but not how they're stored. The idea is to let the user search for
strings using a subset of characters and then figure out what byte
represents each of those characters, assuming that A-Z and a-z are
sequential.

This works for the Pokemon games (for which I already know the encoding
anyway), but fails for FFVA. It stores it's strings some other way,
which will require some more guess work. It stores it's images some
other way too, as there are no LZ compressed images or palettes anywhere
in the ROM.