Bug fixes
Improved keyboard control and fixed some focus issues. Updated the
string decoder to have a concept of stride.
Fixed a bug in pointers that caused it to fail if no pointers were found
before the general "claim the rest" call.
Cursor updates when you jump
added notes on ffva datatypes