Add scripting support
IronRuby scripting is now possible with the app. As the multibox area
was starting to get large, it's been factored out to it's own file.
There's a problem with the commands, so a small hack was added to deal
with that for now: there is likely a way to route the commands from the
window to the user control that I don't know yet.