Initial Revision of the Wild Data
Wild Data is an array of unknown length. We need to find both the array
and its length using the dynamic search system used form Maps. As such,
the search system has been promoted into a helper class, updated to work
with dynamic arrays, and the wild data structure has been added. Now
that the wild data has been correctly found (along with some possible
false positives), we need to test that it works in each rom, and find a
way to distinguish the real data from the false data.