Complete scripted map layout
The rest of the map layout scripting is now working correctly. Added
back the ruby version, as I prefer the lighter syntax and lambdas. Also,
python doesn't work as well for the one-liners.

Given that this is the second time I've switched scripting languages,
I've gone through the trouble of adding both python and ruby. You cannot
currently switch on the fly, but I expect that may become useful later.
A two-line compile time change is currently all I need to switch between
the two languages.