Performance Improvements
Added a concept of a "general layout" when a script is hunting for a
data-type. This allows the Types objects to filter through pointers much
faster, easily discounting ones that don't match the general layout.
This pre-script step massively improves performance, though it does make
the scripts a bit more complicated.