Assertion Fixes
My assertions have been warning me for a while that my datatypes are not
so correct - outside the usual FireRed test, I get several assertions.

- some strings were being read that were REALLY long - like insanely so.
Knowing that strings cannot be above a certain length is helpful.
- runs were careful not to place themselves if another run was located
at their start point; but they didn't check to see if there was a run at
their end point.
- once a destination was added, it was assumed that everything that
*looked* like a pointer to that destination *was* a pointer to that
destination. But just because a destination is found doesn't mean that
all possible pointers actually point to it. Emerald had an example of
this where some signpost data looked like pointers, and the destination
was found before the signpost. Additionally, Ruby had some examples of
things that were assumed to be pointers but conflicted with another
piece of data. Those issues have been fixed now.