First Revision of Builder2
Builder2 is an advancement over Builder. It creates objects that know
the structure of the datatype, but not the value. This means the built
object must lookup the value when requested, which demands the value be
up to date (before it was cached). As an added bonus, this makes it easy
to implement a writer. For a given field (byte/short/int) within a data
type or nested data type, that field can now be written programatically.