dexinfo was getting cut off
there were some char translations missing from the pcs3-w.ini file. To
find them, I formalized a StringFactory capable of providing the
StringDecoder and updated the window to not blap the breadcrumb when
you're reading a different interpretation of the same file.

Also made a few misc edits to scripts to prep for making all the scripts
follow the same conventions.