Remove Evil PCS Singleton
Singletons are evil. There was a hard-to-detect bug that snuck into the
code a while ago that manifests when you try typing a string after
reloading. The PCS ended up with 2 event handler callbacks - one new one
and one old one that never got removed because of its singleton status.
It called up the chain and eventually reached a higher level that _was_
removed, causing the app to come crashing down on itself.

To fix this problem, I removed the singleton and added explicit
dependencies in the places that need the PCS.