first revision of dexorder
dex order feels like it's backwards. It's a list of all species, listing
what order they appear in the dex. This allows to to quickly see, for a
species, where it is in the dex. A more natural way to view the list
would be: for a index in the dex, see what species is there. Eventually
I may want some kind of custom display for this to make it possible to
edit this list without losing your mind.

Now that the dexorder is known, I can use it to list pokemon names as
headers for the dex data.