Cleanup scripting
made all the scripts follow some conventions to hopefully make them
easier to follow. Added a _setup script (named to execute first) and
added some "constants" to it for the other scripts to use. Made the
BuildableArray enumerable and gave it an "each" method so IronRuby could
run for loops on it (for epic).