update epic to evolve pokemon
- moved setup scope. Lazy is "smarter" except that I want app.status
available for debugging purposes on scripts.
- scripting popup is smarter now - doesn't popup with the "last result"
whenever there is no result.
- epic script now updates pokemon species by evolving it to the first
available evolution, twice.
- removed redundant names from epic, as that's a bad coding practice.
- In trying to write the update for epic, I noticed all sorts of trouble
with BuildableArray. I've removed it and replaced it with a simpler
collection of BuildableObject[]. This means that BuildableObject doesn't
need to change locations anymore, so it can expose its location,
removing the need for the strange destinationof method.
- fixed a bug in the dexinfo label: it should fail for any number less
than 1, not any number less than 0.
- With BuildableArray gone, I need a new labeler. ScriptedDataTypes has
absorbed this.
- BuildableObject still needs some way to by dynamic as its data
changes: so it uses a ChildReader and a Builder2 to build arrays each
time they are requested. It does, however, assume static length. Under
the current system, data can change and move, but the structure of the
data once created must remain relatively static.