Added byte enum support
also fixed a bug in the short enum where all attacks would be listed as
attack1 instead of attack 1-4.
Decided that pokemon moves are a more complicated data structure than
scripting would allow, so I'll need to pull the moves out and do them in
code. added PokeMoves to eventually fill that goal.