Research code for non-compressed images.
Without compression as a guide, it's difficult to tell that a block of
data is an image. I have the following hints:
- it must be pointed to.
- it will not have a pointer leading to it.
- images come in 8x8 pixel blocks, or more commonly, 16x16 pixel blocks.
- images store two pixels per byte. As the upper-left pixels in an image
are usually transparent, I can assume images will start with 0x00.

Putting this together, I have an algorithm that finds "likely" images to
display to the user: the user would then be in charge of picking which
are really images and then finding the width/height through trial and
error.

Later, I plan to allow the user to pick which ones are likely images and
then export them to a ruby file for inclusion in the ROM. This editor
needs to be a public addition, but must be turned on manually. It will
likely continue to be in a second window.