Remove scripts from project
While it's important to have the scripts, manually adding them to the
project and changing the output type to "copy always" has become
tedious. It's better to just have a set of scripts in a directory, and
then copy those scripts whole-sale to the output directory where they're
used. Works the same way, but with a simpler configuration and less
repetition.