Fix the find pointer bug
Before this change, an odd pointer (due to lack of alignment of pointer
to a location that is not %4==0) will only recognize the first addition
to it. That is, if multiple sources for that odd pointer are found, only
the first is recognized visually as a pointer. After this change, any
number of sources may be found and added to the list of sources for a
given pointer.