Species Element Provider is searchable
Species Element Provider is no longer a singleton, but we still create
only one instance of it. That instance takes in a name list which
describes the pokemon names available in its enum dropdown list as well
as the pokemon names that should be searchable.

Note that, right now, enum searches require _exact_ name matches
(disregarding case), so "ivy" won't match "ivysaur" even though it would
in a string search. This limits the number of results you get when
searching for things like "char".