Refactor "{ name }" code
There were several places where I was decoding an array of dynamic
objects using a rule. If it's ToString() returns "{ name }" then assume
a name property/field and return that. Otherwise, return the ToString().
This works well, but may need to be adapted in the future. Thus all the
instances have been pulled out into a single utility method for future
editing.