Added BuilderCache
Builder had a lot of static collections that would make life difficult
when reloading or opening a new file. The statics have been moved into a
cache object that is created at load time. The object is anemic right
now, but should gain methods and protection if its code continues to
change / grow.