First pass: replace data locations with segments
not all data should be zero-based: right now, I assume that all data
comes from a single byte array, and I give that array to everyone and
their brother. Everyone also stores an offset location of where the data
they're interested in actually is. The Segment pattern should
encapsulate the array and the offset, so users know longer need to know
how pointers work and that it's all one big array.