Push segment up another level
At this point, the IModel is using a Segment for the entire Model,
passed in at construction time. The Model exposes this Segment so the
view can call things on the window's Holder. This way, even the view can
join in on the fun, even though the original segment is created at the
model level.