initial revision of datatabs
data tabs is a feature that will eventually allow me to store pointers
to multiple file locations, as well as store data not in the main file.
In order for it to work, all the current data must be transfered to an
initial "home" tab. This change begins that adaptation by shifting
properties into the DataTab class and making the other code call that
instead of taking the data directly from the window.