update cursor control
the highlight-when-hover mechanic for the interpretation pane now works
correctly again. Also, after a highlight, keyboard controls move you to
the beginning or end of the highlight instead of ignoring the highlight.

Due to the cursor always jumping when the window jumps, move-to-end
didn't work if the end was not onscreen. Now the cursor stays where it
is, so it's not a problem now.