Add double-click tab opening support
The breadcrumb bar no longer does anything - instead, whenever we would
add something to the breadcrumb, we open a new tab using the current
tabs attributes, then go to that location in the tab.

The tabs need some way to store their head separate from their current
location. They also need some way to click the tab to got to that
location, as well as a way to close any tab except the main tab.