initial duplication
The whole point of adding tabs was so that I could duplicate data and
mess with the duplicate in a new tab. The first step towards duplication
is now available - users can highlight data and duplicate it to a new
tab. Currently the data cannot be push back to the rom and has no
association, but it is a correct copy.