Duplicating data inside a run storage
Now the data inside a run storage is duplicated whenever a part of the
runstorage is duplicated. It's a little messy, but it allows you to keep
formatting whenever you duplicate a segment of data.

Next up is probably making duplicated pointers act correctly (open on
the main data), and making labels copy over as well.