Inclusion of the "+" button
Slowly pushing to make it possible to add new lines to the data. The
basic idea is to add a "+" button at the end of duplicated data that
lets you add new lines, as well as the ability to delete lines in some
way yet to be decided. Also added are interface functions for finding
free space, re-adding duplicated data, and repointing to the new data.
However, these are just the interface functions - the implementations
haven't been added yet.