tab bug fixes
DuplicateTabs should keep their columns constant, but they should indeed
change their rows as the user requests.

There was a bug when adding a new line to a duplicate tab. The new line,
and the add button, had to all be in view, since the model then requests
an update of all those bytes. This is a hint into a larger problem: the
same thing would occur if editing a string when the entire string is not
onscreen. To fix it, the update method has been fixed to ignore
requested updates that are offscreen.