Implementing Repoint
At some point after the space has been found and the data has been
added, the pointers will need to be updated to point to the new
location. This change adds that. To accommodate this and future
operations which require the view to have pointer control, a new "model
operations" interface has been added. It will likely need to be renamed.
This allows the PointerMapper to explain to the CompositeModel how to do
the repointing operation.