find free space and repoint
Basic repointing functionality is added now. Somewhere along the line,
likely with the Segments, I broke the "save" feature. I need to repair
that to see if this is actually working as it should.

The column resizing is also a little flaky now. Need to find out what's
wrong with it.